Attempt to make spec
Attempted to make a spec for the service (issue #32), but was unable to finish at this time.  The next step is to generate a dummy owner with a connected google identity, and from there.  From there, a call can be made to the google api to retrieve the relevant calendar id, which can then simply be displayed in an iframe.  However, setting this up requires some work with google authorization.