Create naive editMessage function
Create a first draft of an editMessage function.  The message is incomplete and tests have not been fully written, as I am unsure of where it is correct to have PUT requests transmit their data.  Additionally, it would make some sense to use a PATCH request instead of PUT, but this would make it more difficult to make expiration date optional in the future.  Although editing messages is not itself a direct issue, it is an unexpected requirement of issue #18.