Make expiration date nullable with tests
In order to be able to determine if a notice is actually intended to expire, the expireDate parameter is now nullable with the expectation that non-expiring notices will have a null expiration date.  Added minor changes so that nullability does not introduce bugs, as well as tests on adding non-expiring notices, and adding or removing expiration dates on existing notices.  Part of issues #17 and #18.